Move AbstractControllerAPI to controller.api package
